requirements.txt
logmatic-python==0.1.7
pandas==0.24.2
sqlalchemy==1.3.3
seaborn==0.9.0
pymssql==2.1.4